Improve storybook examples, rationalize mocks a bit more
