+++ packages/components/src/Drawer/stories/renderProps.story.tsx
@@ -25,11 +25,11 @@
  */
 
 import React from 'react'
-import { SampleContent } from '../../__mocks__/SampleContent'
+import { DialogLongContent } from '../../__mocks__/DialogLongContent'
 import { Drawer } from '../Drawer'
 
 export const RenderProps = () => (
-  <Drawer content={<SampleContent />}>
+  <Drawer content={<DialogLongContent />}>
     {(drawerProps) => <button {...drawerProps}>Open Drawer</button>}
   </Drawer>
 )
