+++ packages/components/src/__mocks__/DialogLongContent.tsx
@@ -24,39 +24,16 @@
 
  */
 
-import React, { useContext } from 'react'
+import React from 'react'
 import styled from 'styled-components'
-import {
-  DialogContext,
-  DialogHeader,
-  DialogContent,
-  DialogFooter,
-  Button,
-  ButtonTransparent,
-  Heading,
-  Paragraph,
-} from '..'
+import { Heading, Paragraph } from '..'
+import { DialogLayout } from './DialogLayout'
 
-export const SampleContent = () => {
-  const { closeModal } = useContext(DialogContext)
-
-  return (
-    <>
-      <DialogHeader hideClose>
-        The Constitution of the United States
-      </DialogHeader>
-      <DialogContent>
-        <Constitution />
-      </DialogContent>
-      <DialogFooter>
-        <Button onClick={closeModal}>Done Reading</Button>
-        <ButtonTransparent color="neutral" onClick={closeModal}>
-          Finish Later
-        </ButtonTransparent>
-      </DialogFooter>
-    </>
-  )
-}
+export const DialogLongContent = () => (
+  <DialogLayout title=" The Constitution of the United States">
+    <Constitution />
+  </DialogLayout>
+)
 
 const Constitution = () => (
   <Format>
